Depsgraph: Rework the way how layer visibility are handled
The main goals of this commit are:

- Get rid of list of IDs which were tagged for update but are not visible
- Make it possible to have proper time dependency re-tags of invisible objects

This change is really quite the same as previous one: just uses ob->recalc
flags to check whether object was updated or not. It might be a bit of an issue
with materials, but they're always "visible", so should not cause any missing
updates.

In any case, we really need to get rid of all this temp sets and arrays, they
might become quite huge in production files.