Depsgraph: Fix tweaking animated custom property doesn't work
It was not totally correct logic around when to tag animation component
for update.

Now its tagging when adt->recalc is set to ADT_RECALC_ANIM, which means
for now setting this flag should happen prior to DAG_id_tag_update() call,
but in the future we can wrap this into an utility function.