Cycles point density: Avoid race condition when using same object as texture source
If the same object is used by two point density textures in Particle source mode
it was really easy to catch a crash in derived render creation.

For now solved with mutex lock around cache_pointdensity_ex() which is not ideal
but seems to be better compromise comparing to other alternatives.