Initial support for user defined Object Widgets
Adds a Panel to the Object Context in the Properties Editor, currently
only containing a Menu to Add widgets. The menu contains 4 items,
Translate, Rotate, Scale and Custom, whereby we currently only draw a
translate widget for all of them. So yes, basic drawing also works.

Important is, that this makes it possible for users to attach custom
widgets to Objects. (Only the manipulator for the selected object is
drawn).