Cycles kernel split: Get rid of loading kernels from run_cpu/run_gpu
It's not proper way to do it and proper approach should be:

- Requested device capabilities should be calculatable prior to device_update.
  This basically means all the things are to be coming from C++ classes, not
  from packed textures and so.

- Kernel might be needed to be reloaded (for example if node set is changed
  in the viewport, but it'll need to have proper check around if kernel really
  need to be reloaded.

- Device capabilities are to be fast to calculate and for the clearity of the
  model should be always passed to the device, without extra checks whether
  device actually needs those capabilities.

Current code is not totally optimal for the viewport, but in terms of node
set it matches previous behavior and it also fixes issues around possible
wrong max closure counter passed to the kernel after tweaks in the shader
tree.

P.S. Capability is getting a bit overused here, would need to make clear
naming for it.