Cycles kernel split: Always gather requested feature set for the device
It is a bit violation of design to check what kernel type device will use
prior gathering the required features. It'll also fail for cases like
network rendering.

This bit of code is relatively cheap, so there's no real harm doing it
always and letting the device itself to decide ignore or use the info.