Cycles kernel split: Remove unneeded changes in device multi
Setting multi device's split kernel flag is a bit misleading and actually
malformed. it might be in theory heterogeneous environment where it'll be
totally unclear how to treat this flag.

Since it's now only used for skipping samples report in the status line
and multi devices are ignored anyway it's safe to remove that bit of code.

TODO: Still have feeling we can get rid of this flag in device info and
just treat reported number of samples as scheduled number of samples.
This could also help in the cases like distributed rendering.