Cycles kernel split: Move device requested features to own struct
Previously it was all stored in the device itself, which is not really good
practice. Now all parameters which are related on what exact kernel with
what feature set to load are wrapped into own struct which is passed to
load_kenrels() function.

Network render part is not really tested and likely is broken.