Cycles kernel split: Fix broken logic around multi-closure in device_opencl
The code actually acted in a way when multi-closure was always enabled, this is
due to kernel_types.h header by default using CPU settings.

This caused re-definition of maximum number of closures.

Now the code acts in a way so desired number of closure is passed via cflags
and then kernel decided whether to use it or ignore.

It's possible that some OpenCL platform still over-allocates, but that's not
caused by this change and needs to be addressed separately.