KTX file format support
This commit introduces support for the KTX format in blender.

Saving and loading of files is supported though saving compressed formats
is not yet supported.

Thumbnails are not generated because the API relies on OpenGL to load the textures,
which is not available during threaded thumbnail generation.

We hand compiled a KTX library from https://github.com/KhronosGroup/KTX
and placed KTX lib files in intern/KTX as follows:

libktx.a in intern/KTX/lib
ktx.h in intern/KTX/include

We may include the library in extern when we make sure it compiles in all platforms (Ubuntu 15.04 failed here)

Scons is not supported, only cmake for now.