Redesign area/region edges
Mainly removed unnecessarily drawn edges and doubled line thickness for
area edges to clearly separate areas from regions.

For now I've added a theme option to control the color of the area
edges, we can check later if it's worth keeping it, but IMHO it's quite
important to have this controllable.