Cycles: Add an option to make object only use direct light
This is a performance trick which could be used for objects which are rather
complex (like grass field) but are out of focus or heavily blurred. For such
objects it does not make sense to do huge number of ray bounces because this
wouldn't give any more details to the result.

Objects might become a bit darker with this option enabled (similar to cases
when number of bounces is limited on scene level), but this is possible to
compensate with material and light settings.

Currently only supported by regular path tracing in megakernel. Before making
it something more widely supported we'd better see what's the best way to
perform such an early ray terminations, how to control them from the UI and
stuff like that.

This seems to give like up to 10% speedup with 01_03_01_B. But need to double
there's no measurable regressions with this feature disabled prior to merge
it to gooseberry branch/