Port Subsurf to VBO code from GPUDataRequest branch.
This adds support for vertex buffers for solid shaded mode for subsurf
modifiers, making drawing of subsurf much faster in this mode.

It also moves towards the goal of data driven requests for our new
renderer system, but this will need further changes down the line.

Everything should work as before with the exception of simplified
multires drawing in sculpting (feature enabled while rotating the view).