Use index buffers for drawing with vertex buffers.
Tests show that we gain approximately 20-25% performance
by that for solid mode drawing.

Good memory/performance is gained for quad meshes mostly
but we also gain performance in triangle meshes as well.

This commit will not get rid of the CPU overhead in
some modes (textured/GLSL) yet, but it prepares the
code for caching changes to make things better
.