Subsurf drawing:
Solid shading uses indexed drawing (textured/GLSL drawing are still not
implemented).

For subsurf we might be able to squeeze more memory but it would only
work for smooth shading and it requires more elaborate counting for all
data upload functions. May be interesting to do in the future though.