Cycles: Implement sample offset feature
This is real quick implementation of the feature which allows to render
different sample ranges on different machines and merge the result into
final image as if it was rendered on a single machine. Same is feature
makes it possible to refine animation with more samples after it was
rendered with lower number of samples.

Control is happening via the following environment variables:

- CYCLES_RENDER_SAMPLE_START, which defines which sample to start
  rendering from.

- CYCLES_RENDER_NUM_SAMPLES, which defines how much samples are to be
  sampled by the Blender instance.

Limitations:

- Not really intuitive, final implementation will require better
  integration into command line arguments.

- All the render layers are to have the same exact number of samples,
  meaning using samples override for layer will lead to undesired
  results.

- No compositor is allowed.