GLSL drawing redesign:
Use new upload scheme. Idea here is that materials share a VBO where
the size of each element is the maximum size of all mesh materials.

While this will waste some size if material sizes differ it is the
simplest scheme to use and allows easy reuse of indices as opposed to
separating the materials in separate vertex buffers. In fact if we do
that, management gets quite complex and code much more error prone.

I may write an extra blog post to explain the choices here
at some point.