OpenSubdiv: Implement volatile the limit evaluator
This evaluator can be used to evaluate the limit from multiple threads.

Not totally optimal -- still need to avoid per-evaluation heap memory
allocation and make all the data stored on stack.

Another optimization will be to make this new evaluator to evaluate
whole bucket of patch coordinates at once in order to avoid API overhead,
but this would require changes in CCGSubSurf.