GP Sculpt: "Grab" Brush
This brush can be used to move clusters of points around, without needing to
select them first. Like the "Grab" brush in Sculpt Mode, this brush only operates
on the brush circle at the start of the grab operation. The strength of the effect
applied to each point is determined at the start of the stroke based on factors
such as how far away from the center of the brush the point lies, and the strength
of the brush.

(Yay! Worked first time!)