GPencil Sculpt: "Push" Brush
This is similar to the grab brush in that it translates points according to
the strength of the brush. The difference however is that it is applied instead
to whatever points are under the brush as it travels, instead of only those
that were within the brush region initially.

Notes:
* The effect of this brush is more like a "smear" effect
* I got a crash initially, but haven't been able to reproduce. So, unless this
  crops up again, I'll consider that a one-off for now.