GP Sculpt: Initial implementation of "Pinch" Brush
This brush draws all the points inwards towards the midpoint of the brush.
I'm not sure if this brush is really that useful or not, so I'll leave it
as-is for now.

Watch out for how strong you set the brush to be; even with 0.3, you can already
get some really strong/rapidly changing results.