OpenSubdiv: Initial wok over separation of legacy and new evaluators
Since the CCGSubSurf.c file started to grow rather insanely and not being
possible at this moment to drop old evaluator it makes sense to split
evaluators into own files. This way code becomes a bit easier to follow,
plus in the future it should help isolating and removing code which is
no longer used.

Still some polishing is possible, but need to get started somewhere.