OpenSubdiv: Initial work related on skipping CCG geometry creation for GPU
The idea is to only allocate CCG geometry (CCGVert, CCGEdge and so on) only
int cases when CPU evaluator is used. In case when GPU rendering is enabled
we'll be creating topology refiner directly from DerivedMesh.

Still quite enough of work to be done, including bringing back GPU backend
synchronization code.