GP Sculpt: Sculpting on a frame without an existing key will create a new one
Just like with the transform tools, if you start trying to sculpt strokes on a frame
where no gpframe/drawings exist, a new gpframe will now be created. This makes it
easier to use the sculpt tools for animating, as all you need to do is to start
animating a sketch is to jump to a frame and start sculpting.

The grab tool in particular works quite well for this sort of workflow. It appears
to be particularly handy when doing breakdowns and/or straight-ahead animation, as
well as for doing smaller tweaks.