New, not so chunky arrow widget drawing
We're now using immediate drawing for arrow widgets instead of using the arrays created using the widget export script.

Although this also includes a thinner arrow generated using the export script, I decided to not use it for now, as the thin lines are drawn a bit ugly. I just #ifdef'ed it out.