Don't offset arrow widget range by property min value
Previously, the min value of the arrow range was set to be at the origin, but figured out in most cases this is actually not wanted (e.g: force field strength widget min value is -200, meaning at strength=0, the widget would be offset 200 pixels; camera focal length widget min value is 1, meaning widget was constantly offset by -1). Instead, the origin should be where the value of the property is 0.

This might still be wanted for some future cases, so leaving old behavior #ifdef'ed out.