Added support for tablet pen and eraser events to be added to keymaps
Added two new event types that will fire only when using a stylus or its eraser.
The use case for this is to allow artists to be able to set up their keymaps
so that they can just draw and erase directly using their tablet without holding
down any keys.