GPencil: Initialise default values for before/after onionskin colours
Newly created layers will now get a pair of default values for the before and after
onion skinning colours. These are currently hardcoded to a green and blue colour
(respectively). Also, "use colours" toggle will be enabled by default, so that these
colours will be applied as soon as you show the onionskins.