GP Sculpt: Experimenting to make sculpt operator non-modal
Sometimes the modal nature of this operator made it cumbersome to operate.
This commit tries to make this better by making changes to the operator + keymap
so that:

* EKEY + LMB-drag (just like for drawing) = Do a single sculpt stroke
  (i.e. it starts immediately, and ends as soon as the stroke ends)

* CTRL + EKEY = Enter Sculpt Mode (as before). Can do multiple strokes this way
  without holding down any keys.

* DKEY + EKEY -> Sculpt = Enter Sculpt Mode (as before)


I'm still on the fence about the Ctrl+EKEY one. Perhaps it's too much of a hassle,
and we just drop this one?