Cycles: Add Support for IES files as textures for light strength
This patch adds support for IES files, a file format that is commonly used to store the directional intensity distribution of light sources.
The new "IES Light" node is supposed to be plugged into the Strength input of the Emission node of the lamp.

Since people generating IES files do not really seem to care about the standard, the parser is flexible enough to accept all test files I have tried.
Some common weirdnesses are distributing values over multiple lines that should go into one line, using commas instead of spaces as delimiters and adding an END marker at the end of the file.

The user interface of the node is similar to the script node, the user can either select an internal Text or load a file.
Internally, IES files are handled similar to Image textures: They are stored by the ImageManager and each unique IES is assigned to one slot.

The local coordinate system of the lamp is used, so that the direction of the light can be changed. For UI reasons, it's usually best to add an area light,
rotate it and then change its type, since especially the point light does not immediately show its local coordinate system in the viewport.

Reviewers: #cycles

Differential Revision: https://developer.blender.org/D1543

Conflicts:
	intern/cycles/app/cycles_xml.cpp
	intern/cycles/kernel/kernel_bake.h
	intern/cycles/kernel/kernel_emission.h
	intern/cycles/kernel/kernel_shader.h
	intern/cycles/kernel/kernel_textures.h
	intern/cycles/kernel/kernel_types.h
	intern/cycles/kernel/osl/osl_services.cpp
	intern/cycles/kernel/svm/svm_types.h
	intern/cycles/render/image.cpp
	intern/cycles/render/image.h
	intern/cycles/render/light.cpp
	source/blender/blenkernel/BKE_node.h
	source/blender/blenkernel/intern/text.c

Conflicts:
	intern/cycles/kernel/kernel_bake.h
	intern/cycles/kernel/kernel_emission.h
	intern/cycles/kernel/kernel_shader.h
	intern/cycles/kernel/kernel_types.h
	source/blender/blenkernel/BKE_node.h

Conflicts:
	intern/cycles/kernel/kernel_shader.h
	intern/cycles/kernel/kernel_types.h
	intern/cycles/kernel/osl/osl_services.cpp
	intern/cycles/render/image.cpp
	intern/cycles/render/image.h
	intern/cycles/render/light.cpp