Fix: The "transform strokes on unkeyed frame creates new frame" feature didn't work on the first frame
As a convenience feature, it is possible to jump to a frame without any existing
GP keyframe, and just use the transform tools to modify the stroke to create a new
keyframe on that frame. However, if the current frame was before the first GP keyframe,
the newly created keyframe would not work (if you scrub away), as it was getting
incorrectly added at the end of the list of keyframes, instead of in-order as expected.