GP Sculpt: Clone Brush - "Use Falloff" = "Smudge"
When the "Use Falloff" option for the Clone brush is enabled, moving the mouse
after pasting some cloned strokes will now allow you to distort the points of
the pasted strokes with non-uniform weights. This can be used for creating
"distortion" effects (similar to using the push brush on these strokes afterwards).
This could be useful as a quick way to get lots of variation to a field of items.