GP Edit: Provide more direct quick access to sculpt brush via D+E pie
Added direct buttons to quickly change the sculpt brush when bringing up the
D+E pie menu for GP Stroke Sculpting. These are only included for the most
useflu of the brushes. All others are currently still visible via the
dropdown on the left (of the pie), though I'm considering removing that
to make things less confusing.